+++ packages/flutter_adaptive_cards_fs/lib/src/cards/elements/text_block.dart
@@ -225,7 +225,6 @@ class AdaptiveTextBlockState extends State<AdaptiveTextBlock>
             color: color,
           );
 
-
     final TextStyle aStyle = TextStyle(
       fontWeight: fontWeight,
       fontSize: fontSize,